requirements.txt
# Core dependencies
PyYAML>=6.0
watchdog>=3.0.0

# Analysis engines (installed separately)
# sudo apt install clamav clamav-daemon yara

# GUI dependencies
Flask==3.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Optional ML dependencies (for future ML features)
# torch>=2.0.0
# scikit-learn>=1.3.0

# Security scanning
bandit>=1.7.0
safety>=2.0.0
